+++ stream/stream_lavf.c
@@ -24,6 +24,7 @@
 #include "common/msg.h"
 #include "common/tags.h"
 #include "common/av_common.h"
+#include "misc/thread_tools.h"
 #include "stream.h"
 #include "options/m_config.h"
 #include "options/m_option.h"
